refactor: modularize page builder state
