+++ packages/ui/src/components/cms/page-builder/state/index.ts
@@ -1,0 +1,46 @@
+import type { HistoryState } from "@acme/types";
+import type { Action } from "./actions";
+import {
+  add,
+  move,
+  remove,
+  duplicate,
+  update,
+  resize,
+  set,
+  undo,
+  redo,
+  setGridCols,
+} from "./reducer";
+
+export type { Action } from "./actions";
+export { historyStateSchema } from "./schema";
+
+export function reducer(state: HistoryState, action: Action): HistoryState {
+  switch (action.type) {
+    case "undo":
+      return undo(state);
+    case "redo":
+      return redo(state);
+    case "set-grid-cols":
+      return setGridCols(state, action);
+    case "add":
+      return add(state, action);
+    case "move":
+      return move(state, action);
+    case "remove":
+      return remove(state, action);
+    case "duplicate":
+      return duplicate(state, action);
+    case "update":
+      return update(state, action);
+    case "resize":
+      return resize(state, action);
+    case "set":
+      return set(state, action);
+    default:
+      return state;
+  }
+}
+
+export default reducer;
